Product validate and Product update

+++ app/Http/Controllers/Admin/ProductController.php
@@ -21,6 +21,21 @@ class ProductController extends Controller
     
     public function store(Request $request)
     {
+        $request->validate([
+            'code' => 'required|string',
+            'barcode' => 'nullable|string',
+            'name' => 'required|string',
+            'category_id' => 'required|numeric',
+            'unit_id' => 'required|numeric',
+            'cost' => 'required|numeric',
+            'price' => 'required|numeric',
+            'quantity' => 'required|numeric',
+            'alert' => 'required|numeric',
+            'image' => 'nullable',
+            'status' => 'required',
+        ]);
+        
+        
         $imageName = null;
         
         if($request->image){
@@ -69,39 +84,62 @@ class ProductController extends Controller
     }
     public function update(Request $request, Product $product)
     {
-            $image_64 = $request->image;
-            $extension = explode('/', explode(':', substr($image_64, 0, strpos($image_64, ';')))[1])[1];  
-            $replace = substr($image_64, 0, strpos($image_64, ',')+1); 
-            // find substring fro replace here eg: data:image/png;base64,
-
-            $image = str_replace($replace, '', $image_64); 
-
-            $image = str_replace(' ', '+', $image); 
- 
-            if($product->image){
-                // Delete the existing image
-                $path = public_path('uploads/products/'.$product->image);
-                if(File::exists($path)){
-                    File::delete($path);
+        $request->validate([
+            'code' => 'required|string',
+            'barcode' => 'nullable|string',
+            'name' => 'required|string',
+            'category_id' => 'required|numeric',
+            'unit_id' => 'required|numeric',
+            'cost' => 'required|numeric',
+            'price' => 'required|numeric',
+            'quantity' => 'required|numeric',
+            'alert' => 'required|numeric',
+            'image' => 'nullable',
+            'status' => 'required',
+        ]);
+        // return $request->all();
+        $imageName = null;
+        if($request->image){
+            if($product->image != $request->image){
+                $image_64 = $request->image;
+                $extension = explode('/', explode(':', substr($image_64, 0, strpos($image_64, ';')))[1])[1];  
+                $replace = substr($image_64, 0, strpos($image_64, ',')+1); 
+                // find substring fro replace here eg: data:image/png;base64,
+                $image = str_replace($replace, '', $image_64); 
+                $image = str_replace(' ', '+', $image); 
+        
+                if($product->image){
+                    // Delete the existing image
+                    $path = public_path('uploads/products/'.$product->image);
+                    if(File::exists($path)){
+                        File::delete($path);
+                    }
                 }
-            }
 
-            $imageName = time().'.'.$extension;
-            $path = public_path('uploads/products/' . $imageName);
-            File::put($path, base64_decode($image));
-            
-            $product->update([
-                'code' => $request->code,
-                'name' => $request->name,
-                'category_id' => $request->category_id,
-                'unit_id' => $request->unit_id,
-                'cost' => $request->cost,
-                'price' => $request->price,
-                'quantity' => $request->quantity,
-                'alert' => $request->alert,
-                'image' => $imageName,
-                'status' => $request->status,
-            ]);
+                $imageName = time().'.'.$extension;
+                $path = public_path('uploads/products/' . $imageName);
+                File::put($path, base64_decode($image));
+            }else{
+                $imageName = $product->image;
+            }
+                
+        }
+
+       
+        
+        $product->update([
+            'code' => $request->code,
+            'name' => $request->name,
+            'category_id' => $request->category_id,
+            'unit_id' => $request->unit_id,
+            'cost' => $request->cost,
+            'price' => $request->price,
+            'quantity' => $request->quantity,
+            'alert' => $request->alert,
+            'image' => $imageName,
+            'status' => $request->status,
+        ]);
+        
         return response()->json(['success' => 'Product updated successfully']);
     }
